+++ tests/testthat/test-join.R
@@ -54,6 +54,15 @@ test_that("semi_join()", {
   expect_identical(index(x), index(semi))
 })
 
+test_that("semi_join() #122", {
+  semi <- pedestrian %>% 
+    semi_join(
+      tibble(Sensor = "Birrarung Marr", Date_Time = Sys.time()),
+      by = "Sensor"
+    )
+  expect_identical(semi, pedestrian %>% filter(Sensor == "Birrarung Marr"))
+})
+
 test_that("anti_join()", {
   anti <- x %>% anti_join(y_key)
   expect_is(anti, "tbl_ts")
@@ -69,4 +78,5 @@ test_that("mutual key and index #102", {
   expect_identical(left_join(x, xx), x)
   expect_named(left_join(x, xx, by = "year"), c("year", "grp.x", "grp.y"))
   expect_named(left_join(x, yy, by = "grp"), c("year.x", "grp", "year.y"))
+  expect_error(left_join(x, yy, by = "year"), "valid")
 })