+++ NEWS.md
@@ -1,3 +1,14 @@
+# tsibble 0.8.1
+
+* **Breaking change**: `index_by(.data, <empty>)` by default groups the index variable rather than previously `ungroup()`.
+* Added new `unnest_tsibble()` to minimise the impact from the upcoming API changes in `tidyr::unnest()`.
+* `index_by()` allows for grouping index variable; and `group_by()` will throw out an error for grouping index.
+* When creating a tsibble, a warning might be issued for mismatch between interval and index representation.
+* Fixed bugs in `semi_join.tbl_ts()` and `anti_join.tbl_ts()` without suffix (#122).
+* Deprecate `as.tsibble()` in favour of `as_tsibble()`.
+* Defunct `fill_na()` in favour of `fill_gaps()`.
+* Defunct `pull_interval()` in favour of `interval_pull()`.
+
 # tsibble 0.8.0
 
 ## Breaking changes